requirements.txt
# Core
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-Admin==1.6.1
Flask-Babel==4.0.0
Flask-Caching==2.1.0
Flask-Compress==1.14
Flask-DebugToolbar==0.14.1
Flask-Limiter==3.5.0
Flask-Marshmallow==1.0.0
Flask-Principal==0.4.0
Flask-Security-Too==5.3.2
Flask-Session==0.5.0
Flask-Uploads==0.2.1
Flask-User==1.0.2.2
gunicorn==21.2.0
python-dotenv==1.0.1
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.28
alembic==1.13.1

# Security
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.1.0.post1
eventlet==0.35.2
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.0
MarkupSafe==2.1.5
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
passlib==1.7.4
PyJWT==2.8.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2024.1
six==1.16.0
speaklater==1.3
SQLAlchemy-Utils==0.41.1
typing_extensions==4.9.0
WTForms==3.1.2

# Performance & Caching
python-memcached==1.59

# Monitoring & Logging
sentry-sdk[flask]==1.40.4
prometheus-flask-exporter==0.22.4
structlog==24.1.0

# Background Tasks
celery==5.3.6
flower==2.0.1

# File Handling
Pillow==10.2.0
python-magic==0.4.27
boto3==1.34.34

# Testing
pytest==7.4.3
coverage==7.3.2
pytest-flask==1.3.0
faker==22.6.0

# Development
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
pre-commit==3.6.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Additional packages
psycopg2-binary==2.9.9
redis==5.0.1
python-slugify==8.0.4
markdown==3.5.2
bleach==6.1.0 